requirements.txt
asgiref==3.5.0
certifi==2021.5.30
charset-normalizer==2.0.4
Django==4.0.3
django-cors-headers==3.11.0
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.2
oauthlib==3.1.1
pip==22.1.1
PyJWT==2.3.0
PySocks==1.7.1
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
setuptools==62.3.2
six==1.16.0
sqlparse==0.4.2
tweepy==3.10.0
tzdata==2021.5
urllib3==1.26.6
wheel==0.37.1
